dws: add configurable PostRun timeout
Problem: when unmounts take a long time to complete, it can be
indicative of a failure or a problem. Therefore, it would be
useful for there to be a configurable PostRun timeout, after which
all nodes that have not unmounted their file systems are drained.

Add a configurable PostRun timeout to coral2_dws.