build: check for libcxi
Problem: flux-coral2 requires the libcxi API to interface with
the Cassini NIC driver for VNI support, but we don't currently
look for it.

Find libcxi using pkg-config.
Find the cxi_service executable (which is packaged with the library).

For developer convenience, configure does not fail if these cannot
be found, unless --with-cray-libcxi is specified.  The TOSS build farm
should specify this for release packages.