jobtap: add slingshot plugin
Problem: VNIs need to be reserved for jobs from a pool
in the FLux system instance.

Add a jobtap plugin that allocates one or more VNIs to a
job, then posts an event to the job eventlog that can be
ready by a prolog script (for CXI creation) and a shell plugin
(for environment setup).