rename cray-pmi-bootstrap jobtap plugin
Problem: the name cray_pals_port_distributor jobtap plugin
is no longer appropriately named because it generates
random secrets as well as ports, and it actually does
this for Cray PMI, not for PALS.

Rename the plugin and the job event to cray-pmi-bootstrap.
Update the rc1 scriptlet that loads the plugin.
Update tests.
Update docs.