dws: save rabbit container logs to the KVS
Problem: developing a rabbit container is extremely difficult,
because there is no way to see what the output of the container is
without viewing the pod logs directly in kubernetes, which is not
workable on LC systems where only privileged users have permission
to access kubernetes.

If there are pods associated with a workflow, pick one at random,
grab the last 50k characters of its log, and save it to the job's KVS
under the `rabbit_container_log` key.