cmd: add flux-slingshot utility
Problem: prolog/epilog/housekeeping have to create/destroy
CXI services as root, but the available tools, e.g.  cxi_service(1)
will not be the most convenient to drive from a shell script.

Add a helper program that can also be built out as a Flux
oriented Slingshot diagnostic if needed.