doc: fix statement about CXI service auth
Problem: the slingshot design doc mistakenly states that
a CXI service ID is presented to the driver along with an
I/O request, but that isn't how RDMA works.

Restate to indicate that the CXI service ID is presented to
the driver when requesting direct access to NIC hardware
resources (implied: which are then used without talking to
the driver).