doc: drop ncores from reservation object
Problem: slingshot docs show the reservation object that
is posted to the job eventlog with an ncores key that
we don't use.

ncores (used for scaling NIC resource limits) is determined
on each node rather than once for the job, since allocations
may not be uniform.

Drop the ncores key from the example.