libutil: factor out eventlog_wait_for()
Problem: code for reading special coral2 events from the
job eventlog will be needed in multiple places.

Add a general function for reading an event, with timeout
and use it in the cray-pals shell plugin.

In the new code, the timeout is not reset upon receipt of
each event; it applies to the full wait time.  Therefore,
increase the cray_pals default event timeout from 10s to 30s.