dws: support `dw_failure_tolerance` attribute
Problem: as mentioned in #364, some users would like the rabbits to
be more tolerant of a few failures. When thousand-node jobs fail due
to a couple of compute nodes that fail to mount, it can be very
obnoxious.

In the dws-jobtap plugin, fetch the `dw_failure_tolerance` attribute
from jobspec and send it along with the `dws.create` RPC.