dws: check for systemd env var during log config
Problem: when running in a container on a Mac, the logs for the
`coral2_dws` script vanish. When logging is configured, a systemd
python library is used if it can be imported. However, it can be
imported in the container on Mac, but it doesn't output any logs
since the script is not actually running as a service.

Add a check for a systemd environment variable before directing all
logs to the journal.