dws-jobtap: re-subscribe to exception events
Problem: when the dws-jobtap plugin is reloaded, as e.g. happens on
restart, the plugin will no longer receive exception event
notifications for jobs that existed before the reload, because the
only call to `flux_jobtap_job_subscribe` happens in the `DEPEND`
state callback. This means jobs can get stuck in prologs and epilogs.

Add a call to `flux_jobtap_job_subscribe` in the RUN callback as well.

Fixes #379.