dws: add method to post node failure event
Problem: when some nodes in a job will not have access to the job's
requested rabbit file system, but the job will proceed regardless,
the job will need to have an event posted indicating what nodes
failed, and the job's workflow will need to have its `forceReady`
field set to `True`. Then the job can proceed and an rc1 script
can take appropriate action with the failed nodes.

Add a method to `WorkflowInfo` objects for checking the number of
failures and then posting an `exception` event for the job and
setting the `forceReady` field on the workflow.