dws: refactor `get_servers` function
Problem: `coral2_dws.py` has a function for fetching all Servers
resources associated with a workflow that have a size > 0. However,
there may be need in the future to filter for servers based on other
conditions.

Make the `get_servers` function take a callable that inspects fields
of a Servers resource and determines whether or not the resource
should be returned.