dws: patch computes resource in DataIn state
Problem: rabbit jobs can now sometimes tolerate the loss of some
rabbit file systems or mounts. However, if a rabbit file system
is not created, mounts of the file system on all nodes attached to
the rabbit will necessarily fail because there is nothing to mount.
That would make the `PreRun` state messy and confusing.

Tell NNF software that it should disregard the mounts on any compute
nodes attached to a rabbit that did not create its file system.
Achieve this by patching the `Computes` resource to remove any
such nodes.