rc: avoid unportable bashism
Problem: 01-rabbit-rc uses a bash [[ ]] test but its
shebang invokes sh, which is not portable.

On debian, where sh -> dash, this fails with

broker.err[0]: rc1.0: /etc/flux/rc1.d/01-rabbit-rc: 3: [[: not found

Admittedly this is a coral2 specific package, but it is helpful
to be able to install it on debian test systems.