dws: rename `rack` vertices to `chassis`
Problem: the 'rack' vertices in JGF generated by `dws2jgf` are
misnamed; the physical object they represent is actually a chassis,
not a rack. This can lead to all kinds of confusion when talking
about the layout of elcap systems, because it can be unclear whether
one is talking about a physical rack or a rack vertex.

Rename `rack` vertices to `chassis`.

Fixes #399.