feat: support for flux operator miniclusters
This feature includes the addition to create a Flux Operator MiniCluster
running across some subset of rabbit nodes given that the rabbit.mpi
directive is defined. By default, setting that to true (or anything)
can use a default container base and interactive mode, and everything
from that can be customized. In addition, we have a "flux hop" command
that is able to take the same metadata, populate the RabbitMPI Job
object, and create the Flux MiniCluster using the same classes/logic
but without requiring the HPE stuff and Workflow. This could be used,
but likely will be for testing or for fun.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>