minicluster: support for cleanup and module
This changeset moves the Flux Operator MiniCluster to be a module,
operator, under flux_k8s. I have also cleaned up the organization
of assets, and better coupled the creation of the MiniCluster with
saving the name / namespace so they do not need to be provided again.
We will need to implement the function to see if a user is allowed
to request a MiniCluster, and work further on adding the additional
securityContext needed for production.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>