cli: add chassis cli plugin for clusters with JGF
Problem: many users have requested the ability to allocate a specific
number of chassis on CORAL2 clusters. Sometimes this is because they
want a specific number of rabbits, but sometimes it is for locality.

Add a CLI plugin that allows a chassis count to be specified on the
command line.