cray_pals: disable eventlog timeout
Problem: large jobs were shown to occasionally fail with
ETIMEDOUT on the cray-slingshot eventlog watch, and cray_pals
has an identical timeout.

The timeout can be satisfied by the cray-pmi-bootstrap event
OR a surpassing event.  Since the surpassing event is guaranteed
by RFC 21, the timeout is not helpful.  Drop it.

Update sharness test.